pyproject.toml
[tool.poetry]
name = "nutritionist-ai"
version = "0.1.0"
description = ""
authors = ["Samuel Sublate <samuelsublate.dev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.2.1"
langchain-community = "^0.3.1"
langchainhub = "^0.1.21"
langgraph = "^0.2.32"
langchain-core = "^0.3.7"
langchain = "^0.3.1"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
python-telegram-bot = "^21.6"
sqlalchemy = "^2.0.36"
schedule = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"